Update Dockerfile
https://buddy.works/guides/how-dockerize-node-application

Copy application to /app directory and install dependencies. If you add the package.json first and run npm install later, Docker won't have to install the dependencies again if you change the package.json file. This results from the way the Docker image is being built (layers and cache), and this is what we should do:
COPY package.json /app
RUN npm install
COPY . /app

Not sure if this is now in good order, seemed valid, not tested yet.